+++ lib/src/analyzers/lint_analyzer/reporters/reporters_list/html/models/file_metrics_report.dart
@@ -1,8 +1,5 @@
 // ignore_for_file: public_member_api_docs
 
-import 'package:meta/meta.dart';
-
-@immutable
 class FileMetricsReport {
   final int averageArgumentsCount;
   final int argumentsCountViolations;