+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * feat: support excludes for a separate anti-pattern.
 * chore: restrict `analyzer` version to `>=2.4.0 <3.2.0`.
 * feat: ignore private members for `check-unused-l10n` command.
+* chore: remove `@immutable` annotation from code.
 
 ## 4.9.1